Moved 8 tests from GeneralWorkflowIT to LaunchWorkflowIT
